refac wmap. Devided into sys and mon sniff
